refactor(core-recurring): split recurring plugin modules
